pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "gpu-stock-monitor"
version = "0.1.0"
description = "GPU stock monitoring bot for French retailers, with Discord alerts"
readme = "README.md"
requires-python = ">=3.12"
license = { text = "MIT" }

dependencies = [
    # HTTP client with HTTP/2 support
    "httpx[http2]>=0.27.0",
    # Async SQLite
    "aiosqlite>=0.20.0",
    # HTML parsing
    "beautifulsoup4>=4.12.3",
    "lxml>=5.2.1",
    # Config parsing
    "pyyaml>=6.0.1",
    # CLI
    "click>=8.1.7",
    # Playwright (optional but listed for browser fallback)
    "playwright>=1.43.0",
    # Async task scheduling
    "anyio>=4.3.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=8.1.1",
    "pytest-asyncio>=0.23.6",
    "pytest-cov>=5.0.0",
    "httpx>=0.27.0",  # for testing
]

[project.scripts]
gpu-monitor = "gpu_monitor.cli:main"

[tool.hatch.build.targets.wheel]
packages = ["src/gpu_monitor"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]
addopts = "-v --tb=short"

[tool.coverage.run]
source = ["src/gpu_monitor"]
branch = true

[tool.coverage.report]
show_missing = true
skip_covered = false

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.ruff.lint]
select = ["E", "F", "W", "I", "UP"]
ignore = ["E501"]
